backend_python/requirements.txt
fastapi==0.104.1
# Upgrade uvicorn for improved Windows reload/subprocess stability
uvicorn[standard]==0.30.6
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
stripe==7.8.0
authlib==1.2.1
itsdangerous==2.1.2
jinja2==3.1.2
aiosmtplib==3.0.1
websockets==12.0
email-validator==2.1.0
# NOTE: passlib 1.7.4 expects bcrypt to expose __about__.__version__,
# which was removed in bcrypt 4.1.x. Pin to 4.0.1 to avoid noisy tracebacks.
bcrypt==4.0.1
PyJWT==2.8.0
minio==7.2.12
